backend-python/requirements.txt
flask
flask-cors
pyautogui
requests
python-dotenv
